Updated contract and deployed

+++ hardhat/contracts/tokens/AirdropToken.sol
@@ -12,7 +12,7 @@ contract AirdropToken is ERC20, Ownable {
     function claim(bytes calldata proof) external {
         (address to, uint8 score, uint256 secret) = abi.decode(proof, (address, uint8, uint256));
         require(secret == 5234198, "Invalid proof");
-        _mint(to, score * 1 ether);
+        _mint(to, uint256(score) * 1 ether);
     }
 
     function mint(address to, uint256 amount) public onlyOwner {
